+++ store-system-common/src/main/java/com/store/system/service/OrderService.java
@@ -11,10 +11,10 @@ import java.util.List;
 public interface OrderService {
 
     public boolean handleAliBarcodeOrder(long passportId, String authCode, int type, String typeInfo,
-                                         String title, String desc, double price) throws Exception;
+                                         String title, String desc, int price) throws Exception;
 
     public boolean handleWxBarcodeOrder(HttpServletRequest request, long passportId, String authCode, int type, String typeInfo,
-                                        String title, String desc, double price, String ip) throws Exception;
+                                        String title, String desc, int price, String ip) throws Exception;
 
     public  Pager getAll(Pager pager, long startTime, long endTime, long personnelid, int status,long uid,String name,int makeStatus,long subid) throws Exception;
